Add OauthModule class for module discovery
